+++ nemo/collections/nlp/models/machine_translation/mt_enc_dec_model.py
@@ -17,7 +17,7 @@ import math
 import random
 from dataclasses import MISSING, dataclass
 from pathlib import Path
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Tuple
 
 import numpy as np
 import torch
@@ -32,7 +32,12 @@ from nemo.collections.common.losses import SmoothedCrossEntropyLoss
 from nemo.collections.common.metrics import Perplexity
 from nemo.collections.common.parts import transformer_weights_init
 from nemo.collections.nlp.data import TranslationDataset
-from nemo.collections.nlp.models.enc_dec_nlp_model import EncDecNLPModel, EncDecNLPModelConfig
+from nemo.collections.nlp.models.enc_dec_nlp_model import (
+    EncDecNLPModel,
+    EncDecNLPModelConfig,
+    OptimConfig,
+    SchedConfig,
+)
 from nemo.collections.nlp.modules.common.transformer import BeamSearchSequenceGenerator
 from nemo.core.classes.common import typecheck
 from nemo.utils import logging
@@ -55,6 +60,23 @@ class TranslationDataConfig:
     drop_last: bool = False
     pin_memory: bool = False
     num_workers: int = 8
+
+
+@dataclass
+class MTSchedConfig(SchedConfig):
+    name: str = 'InverseSquareRootAnnealing'
+    warmup_steps: Optional[int] = None
+    warmup_ratio: float = 0.1
+    last_epoch: int = -1
+
+
+@dataclass
+class MTOptimConfig(OptimConfig):
+    name: str = 'adam'
+    lr: float = 1e-3
+    betas: Tuple[float, float] = (0.9, 0.98)
+    weight_decay: float = 0.0
+    sched: Optional[MTSchedConfig] = None
 
 
 @dataclass
